Update PKESK to the latest version of the crypto refresh
This commit changes the PKESK parsing and serialization methods to the
updated encodings of x25519 and x488 ciphertexts for v3 PKESK. Instead
of encoding the symmetric key algorithm in the ciphertext with padding, the
symmetric key algorithm is now encoded unencrypted.
Further, the commit refactors the checksum handling and enforces AES with v3 PKESK
for the above algorithms.